Add files via upload
The project has been developed using a combination of HTML, CSS, JavaScript, PHP, and MySQL technologies. Specifically, it is a client-side college cafeteria management system. The database underlying the system comprises three key tables: the "Order" table, responsible for storing user orders; the "login_user" table, dedicated to storing user login details; and the "gold" table, which serves to identify users with a gold subscription. The seamless operation of the system is facilitated through the connection of MySQL with PHP.